Add files via upload
Task1- Check if a number is palindrome. (a number for example 16461 that remains the same when its digits are reversed)
Task2-Format phone numbers to XXX/XXX-XXX and display wich operator it belongs to(One,T-Mobile or VIP)
Task3-Finding the longest word in a sentence.
Task4-Find if the sign is negative or positive when mulitplicating 3 integers
Task5-Printing the sequence of numbers from 1 to 30 and determining wich number is even and odd.